Remove psalm from most installations
As Psalm is only used in one single job it can be removed in general
and then reinstalled when it is actually required before it is
executed.

This process allows to keep Psalm in the dev-dependencies so that
constributors still have it available but in CI-runs it needs not
to be installed when it is not executed.

(cherry picked from commit 42c84764455b8f151b132d5b32d07e7bc0bdce5c)